pdf-manager/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
attrs==17.3.0
docopt==0.6.2
python-dateutil==2.6.1
six==1.11.0               # via python-dateutil
unidecode==0.4.21
